Refactored Blog directory.
