+++ client/components/Blog/Main/Main.vue
@@ -1,0 +1,60 @@
+<template>
+  <section class="home">
+    <TheLastPostSection />
+    <LatestPostsSection />
+    <CustomerUniversitySection />
+    <AllPostsSection />
+  </section>
+</template>
+
+<script>
+import { mapActions } from 'vuex'
+import TheLastPostSection from '@/components/Blog/Main/TheLastPostSection'
+import LatestPostsSection from '@/components/Blog/Main/LatestPostsSection'
+import CustomerUniversitySection from '@/components/Blog/Main/CustomerUniversitySection'
+import AllPostsSection from '@/components/Blog/Main/AllPostsSection'
+
+import initLazyLoadMixin from '@/mixins/initLazyLoadMixin'
+
+export default {
+  name: 'Main',
+  components: {
+    TheLastPostSection,
+    LatestPostsSection,
+    CustomerUniversitySection,
+    AllPostsSection,
+  },
+
+  mixins: [initLazyLoadMixin],
+
+  data() {
+    return {
+      pageSize: 12,
+    }
+  },
+
+  created() {
+    this.getContent()
+  },
+
+  methods: {
+    ...mapActions(['getHomePageContent', 'getBlogPosts']),
+
+    getContent() {
+      // Query to get blog home content
+      this.getHomePageContent()
+
+      // Query to get posts content to preview
+      this.getBlogPosts()
+    },
+  },
+}
+</script>
+
+<style lang="scss" scoped>
+@import '../../../assets/styles/_vars';
+
+.home {
+  padding-top: 100px;
+}
+</style>
