+++ client/components/Blog/Post/SlicesBlock.vue
@@ -9,42 +9,42 @@
       <!-- Text slice template -->
       <template v-if="slice.slice_type === 'text'">
         <!-- Here :slice="slice" passes the data to the component -->
-        <text-slice :slice="slice" />
+        <TextSlice :slice="slice" />
       </template>
       <!-- Quote slice template -->
       <template v-else-if="slice.slice_type === 'quote'">
-        <quote-slice :slice="slice" />
+        <QuoteSlice :slice="slice" />
       </template>
       <!-- Image with caption slice template -->
       <template v-else-if="slice.slice_type === 'image_with_caption'">
-        <image-caption-slice :slice="slice" />
+        <ImageCaptionSlice :slice="slice" />
       </template>
       <template v-else-if="slice.slice_type === 'image_with_attributes'">
-        <image-attributes-slice :slice="slice" />
+        <ImageAttributesSlice :slice="slice" />
       </template>
       <template v-else-if="slice.slice_type === 'embed'">
-        <embed-slice :slice="slice" />
+        <EmbedSlice :slice="slice" />
       </template>
       <template v-else-if="slice.slice_type === 'codeblock'">
-        <code-block-slice :slice="slice" />
+        <CodeBlockSlice :slice="slice" />
       </template>
       <template v-else-if="slice.slice_type === 'divider'">
-        <divider />
+        <DividerSlice />
       </template>
       <template v-else-if="slice.slice_type === 'ordered_list'">
-        <ordered-list :slice="slice" />
+        <OrderedList :slice="slice" />
       </template>
       <template v-else-if="slice.slice_type === 'doublecolumn_bordered'">
-        <double-column-bordered-slice :slice="slice" />
+        <DoubleColumnBorderedSlice :slice="slice" />
       </template>
       <template v-else-if="slice.slice_type === 'github_gist'">
-        <github-gist-slice :slice="slice" />
+        <GithubGistSlice :slice="slice" />
       </template>
       <template v-else-if="slice.slice_type === 'image_gallery'">
-        <gallery-slice :slice="slice" />
+        <GallerySlice :slice="slice" />
       </template>
       <template v-else-if="slice.slice_type === 'section_id'">
-        <section-id-slice :slice="slice" />
+        <SectionIdSlice :slice="slice" />
       </template>
     </section>
   </section>
@@ -53,16 +53,16 @@
 <script>
 import CodeBlockSlice from '@/components/Blog/slices/CodeBlockSlice/index.vue'
 import SectionIdSlice from '@/components/Blog/slices/SectionIdSlice'
-import QuoteSlice from '~/components/Blog/slices/QuoteSlice.vue'
-import TextSlice from '~/components/Blog/slices/TextSlice.vue'
-import ImageCaptionSlice from '~/components/Blog/slices/ImageCaptionSlice.vue'
-import EmbedSlice from '~/components/Blog/slices/EmbedSlice.vue'
-import Divider from './slices/DividerSlice'
-import ImageAttributesSlice from './slices/ImageAttributesSlice'
-import OrderedList from './slices/OrderedList'
-import DoubleColumnBorderedSlice from './slices/DoubleColumnBorderedSlice'
-import GithubGistSlice from './slices/GithubGistSlice'
-import GallerySlice from './slices/GallerySlice'
+import QuoteSlice from '@/components/Blog/slices/QuoteSlice.vue'
+import TextSlice from '@/components/Blog/slices/TextSlice.vue'
+import ImageCaptionSlice from '@/components/Blog/slices/ImageCaptionSlice.vue'
+import EmbedSlice from '@/components/Blog/slices/EmbedSlice.vue'
+import DividerSlice from '@/components/Blog/slices/DividerSlice'
+import ImageAttributesSlice from '@/components/Blog/slices/ImageAttributesSlice'
+import OrderedList from '@/components/Blog/slices/OrderedList'
+import DoubleColumnBorderedSlice from '@/components/Blog/slices/DoubleColumnBorderedSlice'
+import GithubGistSlice from '@/components/Blog/slices/GithubGistSlice'
+import GallerySlice from '@/components/Blog/slices/GallerySlice'
 
 export default {
   name: 'SlicesBlock',
@@ -75,7 +75,7 @@ export default {
     TextSlice,
     ImageCaptionSlice,
     EmbedSlice,
-    Divider,
+    DividerSlice,
     GithubGistSlice,
     GallerySlice,
     SectionIdSlice,
