+++ client/components/Blog/skeletons/SkeletonBlogWidget.vue
@@ -25,7 +25,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import '../../assets/styles/_vars';
+@import '../../../assets/styles/_vars';
 
 @keyframes SkeletonLoading {
   0% {
