+++ client/components/Blog/skeletons/SkeletonFeaturedPost.vue
@@ -24,7 +24,7 @@ export default {
 </script>
 
 <style scoped lang="scss">
-@import '../../assets/styles/_vars';
+@import '../../../assets/styles/_vars';
 
 @keyframes SkeletonLoading {
   0% {
