+++ client/components/Blog/slices/QuoteSlice.vue
@@ -5,12 +5,12 @@
       class="block-quotation"
       v-html="$prismic.asHtml(slice.primary.quote)"
     />
-    <text-quote
+    <TextQuote
       v-if="slice.primary.quotation_design === 'brackets'"
       class="post-quote"
     >
       {{ $prismic.asText(slice.primary.quote) }}
-    </text-quote>
+    </TextQuote>
     <div
       v-if="slice.primary.portrait_author.url || authorName"
       :class="`p-${slice.primary.quotation_design}`"
@@ -20,7 +20,7 @@
         v-if="slice.primary.portrait_author.url"
         class="author__image"
       >
-        <prismic-image :field="slice.primary.portrait_author.mobile" />
+        <PrismicImage :field="slice.primary.portrait_author.mobile" />
       </div>
       <div
         v-if="authorName"
