+++ client/components/Blog/slices/SimpleModal.vue
@@ -128,7 +128,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import '../../assets/styles/vars';
+@import '../../../assets/styles/vars';
 
 .modal {
   width: 100%;
